requirements_offline.txt
fastapi
numpy
filterpy
lap
uvicorn
opencv-python
Pillow
torch
torchvision
yolov5
pandas
requests